start with composite refactor
