+++ src/main/java/clickup/api/Authentication.java
@@ -1,6 +1,16 @@
+/*
+ * Copyright (c) 2019 Jalasoft.
+ *
+ * This software is the confidential and proprietary information of Jalasoft.
+ * ("Confidential Information"). You shall not
+ * disclose such Confidential Information and shall use it only in
+ * accordance with the terms of the license agreement you entered into
+ * with Jalasoft.
+ */
+
 package clickup.api;
 
-import clickup.ui.entities.User;
+import clickup.entities.User;
 import core.utils.CredentialDeserializer;
 import core.utils.PropertyReader;
 import io.restassured.builder.RequestSpecBuilder;
