+++ src/main/java/clickup/ui/components/SpaceBar.java
@@ -10,13 +10,5 @@
 
 package clickup.ui.components;
 
-import clickup.ui.BasePage;
-
-/**
- * TopMenu class.
- *
- * @author Maday Alcala
- * @version 1.0
- */
-public class TopMenu extends BasePage {
+public class SpaceBar {
 }
