+++ src/main/java/core/selenium/webdrivers/Firefox.java
@@ -1,3 +1,13 @@
+/*
+ * Copyright (c) 2019 Jalasoft.
+ *
+ * This software is the confidential and proprietary information of Jalasoft.
+ * ("Confidential Information"). You shall not
+ * disclose such Confidential Information and shall use it only in
+ * accordance with the terms of the license agreement you entered into
+ * with Jalasoft.
+ */
+
 package core.selenium.webdrivers;
 
 import io.github.bonigarcia.wdm.FirefoxDriverManager;
