+++ src/test/java/hook/SideMenuHook.java
@@ -1,6 +1,16 @@
+/*
+ * Copyright (c) 2019 Jalasoft.
+ *
+ * This software is the confidential and proprietary information of Jalasoft.
+ * ("Confidential Information"). You shall not
+ * disclose such Confidential Information and shall use it only in
+ * accordance with the terms of the license agreement you entered into
+ * with Jalasoft.
+ */
+
 package hook;
 
-import clickup.ui.pages.SpaceMenu;
+import clickup.ui.components.SpacePanel;
 import cucumber.api.java.After;
 
 /**
@@ -16,7 +26,7 @@ public class SideMenuHook {
      */
     @After(order = 1, value = "@logout")
     public void logout() {
-        SpaceMenu starPage = new SpaceMenu();
+        SpacePanel starPage = new SpacePanel();
         starPage.getSideMenu().logOut();
     }
 }
